Add a way to toggle the display of files in the tree.
Toggling the files leaves only the directory tree for quicker navigation
to the desired folder, where the files can be toggled back on for
selection.

I also improved the way directories were read. Instead of using system
calls to do an `ls` or `DIR` command, I switched to the native Vim
function `readdir()`. This is safer in that it eliminates the chance of
code injection. The MinTree variables that define the OS commands for
reading a directory are gone now too, and the code is a bit shorter.