Make `f` and `F` more robust. Add `;` and `,` commands to complement.
The search() function always looks in folded lines, so I switched to an
execute call with the `/` command and set 'foldopen-=search`. It still
finds and goes to a folded line, but only the first one in the fold. I
think I'll go with this for a while to see if I like it. I can imagine
this scenario: "I know this node is in the list, but it's in a directory
I closed a while back. Take me there, `f`!" If I end up not liking this,
I'll have to manually check whether a found line is folded or not.