Bug fixes and adding `f` command. (#24)
* Tweak the help text for a couple commands.

* Remove the J, K, <C-J>, and <C-K> keys.

* Updated ReadMe with <kbd> tags where appropriate.

* Fix typo in README, shorten the description of the `dm` command.

* Put J, K, <C-J>, and <C-K> back in.

* Make `w` work only on files.

* Fix `cd` when on a file node. I had the wrong modifier.

* Remove the `F` command. I'm going to use that key for something else.

* Add vifm's `f` and `F` keys to find nodes by their 1st character.

* Update README.md

* Cleanup on aisle 3. Make a quicker exit from Wipeout function.

* Make `f` and `F` more robust. Add `;` and `,` commands to complement.

The search() function always looks in folded lines, so I switched to an
execute call with the `/` command and set 'foldopen-=search`. It still
finds and goes to a folded line, but only the first one in the fold. I
think I'll go with this for a while to see if I like it. I can imagine
this scenario: "I know this node is in the list, but it's in a directory
I closed a while back. Take me there, `f`!" If I end up not liking this,
I'll have to manually check whether a found line is folded or not.

* Update README and quickhelp text for f, F, ; and ,.

* Fix quickhelp description of ; and , commands.